requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
python-multipart>=0.0.20,<1
tensorflow>=2.21,<2.22; python_version < "3.14"
pandas>=3.0,<4
numpy>=2.4,<3
scikit-learn>=1.8,<2
joblib>=1.5,<2
Pillow>=12,<13
statsmodels>=0.14.6,<1
